Extract router and break it down on smaller files
