+++ ios/MullvadVPN/Routing/ApplicationRouter.swift
@@ -11,364 +11,6 @@ import MullvadLogging
 import UIKit
 
 /**
- Formal protocol describing a group of routes.
- */
-protocol AppRouteGroupProtocol: Comparable, Equatable, Hashable {
-    /**
-     Returns `true` if group is presented modally, otherwise `false` if group is a part of root view
-     controller.
-     */
-    var isModal: Bool { get }
-
-    /**
-     Defines a modal level that's used for restricting modal presentation.
-
-     A group with higher modal level can be presented above a group with lower level but not the other way around. For example, if a modal group is represented by
-     `UIAlertController`, it should have the highest level (i.e `Int.max`) to prevent other modals from being presented above it, however it enables alert
-     controller to be presented above any other modal.
-     */
-    var modalLevel: Int { get }
-}
-
-/**
- Default implementation of `Comparable` for `AppRouteGroupProtocol` which compares `modalLevel` of both sides.
- */
-extension AppRouteGroupProtocol {
-    static func < (lhs: Self, rhs: Self) -> Bool {
-        lhs.modalLevel < rhs.modalLevel
-    }
-}
-
-/**
- Formal protocol describing a single route.
- */
-protocol AppRouteProtocol: Equatable, Hashable {
-    associatedtype RouteGroupType: AppRouteGroupProtocol
-
-    /**
-     Returns `true` when only one route of a kind can be displayed.
-     */
-    var isExclusive: Bool { get }
-
-    /**
-     Returns `true` if the route supports sub-navigation.
-     */
-    var supportsSubNavigation: Bool { get }
-
-    /**
-     Navigation group to which the route belongs to.
-     */
-    var routeGroup: RouteGroupType { get }
-}
-
-/**
- Enum type describing groups of routes. Each group is a modal layer with horizontal navigation
- inside with exception where primary navigation is a part of root controller on iPhone.
- */
-enum AppRouteGroup: AppRouteGroupProtocol {
-    /**
-     Primary horizontal navigation group.
-     */
-    case primary
-
-    /**
-     Select location group.
-     */
-    case selectLocation
-
-    /**
-     Account group.
-     */
-    case account
-
-    /**
-     Settings group.
-     */
-    case settings
-
-    /**
-     Changelog group.
-     */
-    case changelog
-
-    var isModal: Bool {
-        switch self {
-        case .primary:
-            return UIDevice.current.userInterfaceIdiom == .pad
-
-        case .selectLocation, .account, .settings, .changelog:
-            return true
-        }
-    }
-
-    var modalLevel: Int {
-        switch self {
-        case .primary:
-            return 0
-        case .settings, .account, .selectLocation, .changelog:
-            return 1
-        }
-    }
-}
-
-/**
- Enum type describing primary application routes.
- */
-enum AppRoute: AppRouteProtocol {
-    /**
-     Account route.
-     */
-    case account
-
-    /**
-     Settings route. Contains sub-route to display.
-     */
-    case settings(SettingsNavigationRoute?)
-
-    /**
-     Select location route.
-     */
-    case selectLocation
-
-    /**
-     Changelog route.
-     */
-    case changelog
-
-    /**
-     Routes that are part of primary horizontal navigation group.
-     */
-    case tos, login, main, revoked, outOfTime, welcome, setupAccountCompleted
-
-    var isExclusive: Bool {
-        switch self {
-        case .selectLocation, .account, .settings, .changelog:
-            return true
-        default:
-            return false
-        }
-    }
-
-    var supportsSubNavigation: Bool {
-        if case .settings = self {
-            return true
-        } else {
-            return false
-        }
-    }
-
-    var routeGroup: AppRouteGroup {
-        switch self {
-        case .tos, .login, .main, .revoked, .outOfTime, .welcome, .setupAccountCompleted:
-            return .primary
-        case .changelog:
-            return .changelog
-        case .selectLocation:
-            return .selectLocation
-        case .account:
-            return .account
-        case .settings:
-            return .settings
-        }
-    }
-}
-
-/**
- Struct describing a routing request for presentation or dismissal.
- */
-struct PendingRoute<RouteType: AppRouteProtocol>: Equatable {
-    var operation: RouteOperation<RouteType>
-    var animated: Bool
-}
-
-/**
- Enum type describing an attempt to fulfill the route presentation request.
- **/
-enum PendingPresentationResult {
-    /**
-     Successfully presented the route.
-     */
-    case success
-
-    /**
-     The request to present this route should be dropped.
-     */
-    case drop
-
-    /**
-     The request to present this route cannot be fulfilled because the modal context does not allow
-     for that.
-
-     For example, on iPad, primary context cannot be presented above settings, because it enables
-     access to settings by making the settings cog accessible via custom presentation controller.
-     In such case the router will attempt to fulfill other requests in hope that perhaps settings
-     can be dismissed first before getting back to that request.
-     */
-    case blockedByModalContext
-}
-
-/**
- Enum type describing an attempt to fulfill the route dismissal request.
- */
-enum PendingDismissalResult {
-    /**
-     Successfully dismissed the route.
-     */
-    case success
-
-    /**
-     The request to present this route should be dropped.
-     */
-    case drop
-
-    /**
-     The route cannot be dismissed immediately because it's blocked by another modal presented
-     above.
-
-     The router will attempt to fulfill other requests first in hope to unblock the route by
-     dismissing the modal above before getting back to that request.
-     */
-    case blockedByModalAbove
-}
-
-/**
- Enum describing operation over the route.
- */
-enum RouteOperation<RouteType: AppRouteProtocol>: Equatable {
-    /**
-     Present route.
-     */
-    case present(RouteType)
-
-    /**
-     Dismiss route.
-     */
-    case dismiss(DismissMatch<RouteType>)
-
-    /**
-     Returns a group of affected routes.
-     */
-    var routeGroup: RouteType.RouteGroupType {
-        switch self {
-        case let .present(route):
-            return route.routeGroup
-        case let .dismiss(dismissMatch):
-            return dismissMatch.routeGroup
-        }
-    }
-}
-
-/**
- Enum type describing a single route or a group of routes requested to be dismissed.
- */
-enum DismissMatch<RouteType: AppRouteProtocol>: Equatable {
-    case group(RouteType.RouteGroupType)
-    case singleRoute(RouteType)
-
-    /**
-     Returns a group of affected routes.
-     */
-    var routeGroup: RouteType.RouteGroupType {
-        switch self {
-        case let .group(group):
-            return group
-        case let .singleRoute(route):
-            return route.routeGroup
-        }
-    }
-}
-
-/**
- Struct describing presented route.
- */
-struct PresentedRoute<RouteType: AppRouteProtocol>: Equatable {
-    var route: RouteType
-    var coordinator: Coordinator
-}
-
-/**
- Struct holding information used by delegate to perform dismissal of the route(s) in subject.
- */
-struct RouteDismissalContext<RouteType: AppRouteProtocol> {
-    /**
-     Specific routes that are being dismissed.
-     */
-    var dismissedRoutes: [PresentedRoute<RouteType>]
-
-    /**
-     Whether the entire group is being dismissed.
-     */
-    var isClosing: Bool
-
-    /**
-     Whether transition is animated.
-     */
-    var isAnimated: Bool
-}
-
-/**
- Struct holding information used by delegate to perform sub-navigation of the route in subject.
- */
-struct RouteSubnavigationContext<RouteType: AppRouteProtocol> {
-    var presentedRoute: PresentedRoute<RouteType>
-    var route: RouteType
-    var isAnimated: Bool
-}
-
-/**
- Application router delegate
- */
-protocol ApplicationRouterDelegate<RouteType>: AnyObject {
-    associatedtype RouteType: AppRouteProtocol
-
-    /**
-     Delegate should present the route and pass corresponding `Coordinator` upon completion.
-     */
-    func applicationRouter(
-        _ router: ApplicationRouter<RouteType>,
-        route: RouteType,
-        animated: Bool,
-        completion: @escaping (Coordinator) -> Void
-    )
-
-    /**
-     Delegate should dismiss the route.
-     */
-    func applicationRouter(
-        _ router: ApplicationRouter<RouteType>,
-        dismissWithContext context: RouteDismissalContext<RouteType>,
-        completion: @escaping () -> Void
-    )
-
-    /**
-     Delegate may reconsider if route presentation is still needed.
-
-     Return `true` to proceed with presenation, otherwise `false` to prevent it.
-     */
-    func applicationRouter(_ router: ApplicationRouter<RouteType>, shouldPresent route: RouteType) -> Bool
-
-    /**
-     Delegate may reconsider if route dismissal should be done.
-
-     Return `true` to proceed with dismissal, otherwise `false` to prevent it.
-     */
-    func applicationRouter(
-        _ router: ApplicationRouter<RouteType>,
-        shouldDismissWithContext context: RouteDismissalContext<RouteType>
-    ) -> Bool
-
-    /**
-     Delegate should handle sub-navigation for routes supporting it then call completion to tell
-     router when it's done.
-     */
-    func applicationRouter(
-        _ router: ApplicationRouter<RouteType>,
-        handleSubNavigationWithContext context: RouteSubnavigationContext<RouteType>,
-        completion: @escaping () -> Void
-    )
-}
-
-/**
  Main application router.
  */
 final class ApplicationRouter<RouteType: AppRouteProtocol> {
